ci: run commitlint to check commits (#1602)
As first step in our main "check" workflows (push, linux, mac,
windows), check all the commit messages with commitlint.

These workflows can be triggered in various ways:
pull requests, pushes, manually, or scheduled.

For (each push to) a pull request, all commits currently in the PR
branch are checked.

For a regular push, all the pushed commits are checked.

For a manual run, it seemed to check the same commits as a push (which
push ? Not sure how this works).

For a scheduled run - we'll see.